+++ CHANGELOG.md
@@ -1,5 +1,6 @@
 
 # v4.9.0
+- Added the ability to change the text speed using the built-in settings screen.
 - When debug mode is enabled, `(debug)` is now added to the window's title. This makes it easier to spot if debug mode is enabled/disabled.
 
 # v4.8.1